Move omniauth form partial to a component
This way we simplify the view a little bit and replace some slow system
tests with faster component tests.
