+++ app/views/devise/sessions/new.html.erb
@@ -1,7 +1,7 @@
 <% provide :title, t("devise_views.sessions.new.title") %>
 <h2><%= t("devise_views.sessions.new.title") %></h2>
 
-<%= render "devise/omniauth_form", action: "sign_in" %>
+<%= render Devise::OmniauthFormComponent.new("sign_in") %>
 
 <p>
   <%= sanitize(t("devise_views.shared.links.signup",
