+++ app/views/users/registrations/new.html.erb
@@ -1,7 +1,7 @@
 <% provide :title, t("devise_views.users.registrations.new.title") %>
 <h2><%= t("devise_views.users.registrations.new.title") %></h2>
 
-<%= render "devise/omniauth_form", action: "sign_up" %>
+<%= render Devise::OmniauthFormComponent.new("sign_up") %>
 
 <%= form_for(resource, as: resource_name, url: registration_path(resource_name)) do |f| %>
   <%= render "shared/errors", resource: resource %>
